# Conflicts:
#	src/backend/ci/core/environment/api-environment/src/main/kotlin/com/tencent/devops/environment/api/thirdpartyagent/UserThirdPartyAgentResource.kt
#	src/backend/ci/core/environment/api-environment/src/main/kotlin/com/tencent/devops/environment/constant/Constants.kt
#	src/backend/ci/core/environment/biz-environment/src/main/kotlin/com/tencent/devops/environment/dao/thirdpartyagent/AgentBatchInstallTokenDao.kt
#	src/backend/ci/core/environment/biz-environment/src/main/kotlin/com/tencent/devops/environment/resources/thirdPartyAgent/ExternalThirdPartyAgentResourceImpl.kt
#	src/backend/ci/core/environment/biz-environment/src/main/kotlin/com/tencent/devops/environment/resources/thirdPartyAgent/UserThirdPartyAgentResourceImpl.kt
#	src/backend/ci/core/environment/biz-environment/src/main/kotlin/com/tencent/devops/environment/service/thirdpartyagent/BatchInstallAgentService.kt
#	support-files/sql/1001_ci_environment_ddl_mysql.sql

+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/service/process/ProcessDataClearService.kt
@@ -27,6 +27,7 @@
 
 package com.tencent.devops.misc.service.process
 
+import com.tencent.devops.common.db.pojo.ARCHIVE_SHARDING_DSL_CONTEXT
 import com.tencent.devops.common.db.utils.JooqUtils
 import com.tencent.devops.common.redis.RedisOperation
 import com.tencent.devops.misc.dao.process.ProcessDao
@@ -35,11 +36,14 @@ import com.tencent.devops.misc.lock.PipelineVersionLock
 import org.jooq.DSLContext
 import org.jooq.impl.DSL
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Service
 
 @Service
 class ProcessDataClearService @Autowired constructor(
     private val dslContext: DSLContext,
+    @Qualifier(ARCHIVE_SHARDING_DSL_CONTEXT)
+    private val archiveShardingDslContext: DSLContext,
     private val processDao: ProcessDao,
     private val processDataClearDao: ProcessDataClearDao,
     private val redisOperation: RedisOperation,
@@ -50,35 +54,49 @@ class ProcessDataClearService @Autowired constructor(
      * 清除流水线数据
      * @param projectId 项目ID
      * @param pipelineId 流水线ID
+     * @param archiveFlag 归档标识
      */
     fun clearPipelineData(
         projectId: String,
-        pipelineId: String
+        pipelineId: String,
+        archiveFlag: Boolean? = null
     ) {
-        dslContext.transaction { t ->
+        val finalDslContext = generateFinalDslContext(archiveFlag)
+        finalDslContext.transaction { t ->
             val context = DSL.using(t)
             processDataClearDao.deletePipelineLabelByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineModelTaskByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineRemoteAuthByPipelineId(context, projectId, pipelineId)
             processDataClearDao.deletePipelineResourceByPipelineId(context, projectId, pipelineId)
             processDataClearDao.deletePipelineResourceVersionByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineSettingByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineSettingVersionByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineTimerByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineWebhookByPipelineId(context, projectId, pipelineId)
             processDataClearDao.deleteTemplatePipelineByPipelineId(context, projectId, pipelineId)
             processDataClearDao.deletePipelineBuildSummaryByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineTemplateAcrossInfo(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineViewGroup(context, projectId, pipelineId)
-            // 添加删除记录，插入要实现幂等
-            processDao.addPipelineDataClear(
-                dslContext = context,
-                projectId = projectId,
-                pipelineId = pipelineId
-            )
+            if (archiveFlag != true) {
+                processDataClearDao.deletePipelineModelTaskByPipelineId(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineRemoteAuthByPipelineId(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineSettingByPipelineId(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineSettingVersionByPipelineId(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineTimerByPipelineId(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineWebhookByPipelineId(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineTemplateAcrossInfo(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineViewGroup(context, projectId, pipelineId)
+                // 添加删除记录，插入要实现幂等
+                processDao.addPipelineDataClear(
+                    dslContext = context,
+                    projectId = projectId,
+                    pipelineId = pipelineId
+                )
+            }
             processDataClearDao.deletePipelineInfoByPipelineId(context, projectId, pipelineId)
         }
         processRelatedPlatformDataClearService.cleanBuildData(projectId, pipelineId)
+    }
+
+    private fun generateFinalDslContext(archiveFlag: Boolean?): DSLContext {
+        val finalDslContext = if (archiveFlag == true) {
+            archiveShardingDslContext
+        } else {
+            dslContext
+        }
+        return finalDslContext
     }
 
     /**
@@ -101,14 +119,26 @@ class ProcessDataClearService @Autowired constructor(
      * @param projectId 项目ID
      * @param pipelineId 流水线ID
      * @param buildId 构建ID
+     * @param archiveFlag 归档标识
      */
     fun clearOtherBuildData(
         projectId: String,
         pipelineId: String,
-        buildId: String
+        buildId: String,
+        archiveFlag: Boolean? = null
     ) {
-        dslContext.transaction { t ->
+        val finalDslContext = generateFinalDslContext(archiveFlag)
+        finalDslContext.transaction { t ->
             val context = DSL.using(t)
+            processDataClearDao.deleteBuildRecordPipelineByBuildId(context, projectId, buildId)
+            processDataClearDao.deleteBuildRecordStageByBuildId(context, projectId, buildId)
+            processDataClearDao.deleteBuildRecordContainerByBuildId(context, projectId, buildId)
+            processDataClearDao.deleteBuildRecordTaskByBuildId(context, projectId, buildId)
+            if (archiveFlag == true) {
+                processDataClearDao.deleteBuildHistoryByBuildId(context, projectId, buildId)
+                // 归档库构建数据清理无需执行方法后的逻辑
+                return@transaction
+            }
             processDataClearDao.deleteBuildDetailByBuildId(context, projectId, buildId)
             processDataClearDao.deleteReportByBuildId(
                 dslContext = context,
@@ -120,11 +150,6 @@ class ProcessDataClearService @Autowired constructor(
                 processDataClearDao.deletePipelineBuildTemplateAcrossInfo(context, projectId, pipelineId, buildId)
             }
             processDataClearDao.deleteBuildWebhookParameter(context, projectId, buildId)
-            processDataClearDao.deleteBuildCommits(context, projectId, buildId)
-            processDataClearDao.deleteBuildRecordPipelineByBuildId(context, projectId, buildId)
-            processDataClearDao.deleteBuildRecordStageByBuildId(context, projectId, buildId)
-            processDataClearDao.deleteBuildRecordContainerByBuildId(context, projectId, buildId)
-            processDataClearDao.deleteBuildRecordTaskByBuildId(context, projectId, buildId)
             // 添加删除记录，插入要实现幂等
             processDao.addBuildHisDataClear(
                 dslContext = context,
@@ -141,7 +166,11 @@ class ProcessDataClearService @Autowired constructor(
             val pipelineVersionLock = PipelineVersionLock(redisOperation, pipelineId, version)
             try {
                 pipelineVersionLock.lock()
-                processDataClearDao.deleteBuildHistoryByBuildId(context, projectId, buildId)
+                val deleteResult = processDataClearDao.deleteBuildHistoryByBuildId(context, projectId, buildId)
+                if (deleteResult == 0) {
+                    // 如果删除的记录数为0则无需执行后面的逻辑
+                    return@transaction
+                }
                 // 查询流水线版本记录
                 val pipelineVersionInfo = processDao.getPipelineVersionSimple(
                     dslContext = context,
@@ -150,7 +179,7 @@ class ProcessDataClearService @Autowired constructor(
                     version = version
                 )
                 var referCount = pipelineVersionInfo?.referCount
-                referCount = if (referCount == null) {
+                referCount = if (referCount == null || referCount < 0) {
                     // 兼容老数据缺少关联构建记录的情况，全量统计关联数据数量
                     processDao.countBuildNumByVersion(
                         dslContext = context,
@@ -181,13 +210,15 @@ class ProcessDataClearService @Autowired constructor(
      * 清除流水线被跳过的任务数据
      * @param projectId 项目ID
      * @param buildId 构建ID
+     * @param archiveFlag 归档标识
      */
     fun clearSkipRecordTaskData(
         projectId: String,
-        buildId: String
+        buildId: String,
+        archiveFlag: Boolean? = null
     ) {
         processDataClearDao.deleteBuildRecordTaskByBuildId(
-            dslContext = dslContext,
+            dslContext = generateFinalDslContext(archiveFlag),
             projectId = projectId,
             buildId = buildId,
             skipTaskDeleteFlag = true
